Get rid of ambiguous longs from CrawlHashTable and a lot of misc places.

+++ crawl-ref/source/misc.h
@@ -27,7 +27,7 @@ void turn_corpse_into_skeleton_and_chunks(item_def &item);
 void init_stack_blood_potions( item_def &stack, int age = -1 );
 void maybe_coagulate_blood_potions_floor( int obj );
 bool maybe_coagulate_blood_potions_inv( item_def &blood );
-long remove_oldest_blood_potion( item_def &stack );
+int remove_oldest_blood_potion( item_def &stack );
 void remove_newest_blood_potion( item_def &stack, int quant = -1 );
 void merge_blood_potion_stacks(item_def &source, item_def &dest, int quant);
